+++ danmaku2ass.h
@@ -2,6 +2,6 @@
 #ifndef bilibili_danmaku2ass_h
 #define bilibili_danmaku2ass_h
 
-void danmaku2ass(const char *infile,const char *outfile,int width,int height,const char *font,float fontsize,float alpha,float duration_marquee,float duration_still);
+void danmaku2ass(const char *infile,const char *outfile,int width,int height,const char *font,int fontsize,double alpha,int duration_marquee,int duration_still);
 
 #endif
